Update views.py
feat: Add endpoint for rides with user as driver

Desc: The endpoint receives a POST request with the "departure date", additional information and the user as the driver. From the data, the endpoint creates a ride dictionary with the user as the driver.

Pending: Understand the serializer in order to validate and save the data as a new ride instance in the database.